+++ etna/datasets/datasets_generation.py
@@ -1,19 +1,38 @@
 from typing import List
 from typing import Optional
+from typing import Sequence
+from typing import Union
 
 import numpy as np
 import pandas as pd
 from numpy.random import RandomState
 from statsmodels.tsa.arima_process import arma_generate_sample
 
+from etna.datasets.utils import TimestampType
+from etna.datasets.utils import _check_timestamp_param
+
+
+def _create_timestamp(
+    start_time: Optional[Union[str, TimestampType]], freq: Optional[str], periods: int
+) -> Sequence[TimestampType]:
+    start_time = _check_timestamp_param(param=start_time, param_name="start_time", freq=freq)
+    if freq is None:
+        if start_time is None:
+            start_time = 0
+        return np.arange(start_time, start_time + periods)  # type: ignore
+    else:
+        if start_time is None:
+            start_time = pd.Timestamp("2000-01-01")
+        return pd.date_range(start=start_time, freq=freq, periods=periods)
+
 
 def generate_ar_df(
     periods: int,
-    start_time: str,
+    start_time: Optional[Union[str, TimestampType]] = None,
     ar_coef: Optional[list] = None,
     sigma: float = 1,
     n_segments: int = 1,
-    freq: str = "1D",
+    freq: Optional[str] = "D",
     random_seed: int = 1,
 ) -> pd.DataFrame:
     """
@@ -35,6 +54,11 @@ def generate_ar_df(
         pandas frequency string for :py:func:`pandas.date_range` that is used to generate timestamp
     random_seed:
         random seed
+
+    Raises
+    ------
+    ValueError:
+        Non-integer timestamp parameter is used for integer-indexed timestamp.
     """
     if ar_coef is None:
         ar_coef = [1]
@@ -44,18 +68,18 @@ def generate_ar_df(
         ar=ar_coef, ma=[1], nsample=(n_segments, periods), axis=1, distrvs=random_sampler, scale=sigma
     )
     df = pd.DataFrame(data=ar_samples.T, columns=[f"segment_{i}" for i in range(n_segments)])
-    df["timestamp"] = pd.date_range(start=start_time, freq=freq, periods=periods)
+    df["timestamp"] = _create_timestamp(start_time=start_time, freq=freq, periods=periods)
     df = df.melt(id_vars=["timestamp"], value_name="target", var_name="segment")
     return df
 
 
 def generate_periodic_df(
     periods: int,
-    start_time: str,
+    start_time: Optional[Union[str, TimestampType]] = None,
     scale: float = 10,
     period: int = 1,
     n_segments: int = 1,
-    freq: str = "1D",
+    freq: Optional[str] = "D",
     add_noise: bool = False,
     sigma: float = 1,
     random_seed: int = 1,
@@ -83,6 +107,11 @@ def generate_periodic_df(
         scale of added noise
     random_seed:
         random seed
+
+    Raises
+    ------
+    ValueError:
+        Non-integer timestamp parameter is used for integer-indexed timestamp.
     """
     samples = RandomState(seed=random_seed).randint(int(scale), size=(n_segments, period))
     patterns = [list(ar) for ar in samples]
@@ -100,10 +129,10 @@ def generate_periodic_df(
 
 def generate_const_df(
     periods: int,
-    start_time: str,
-    scale: float,
+    start_time: Optional[Union[str, TimestampType]] = None,
+    scale: float = 10,
     n_segments: int = 1,
-    freq: str = "1D",
+    freq: Optional[str] = "D",
     add_noise: bool = False,
     sigma: float = 1,
     random_seed: int = 1,
@@ -119,8 +148,6 @@ def generate_const_df(
         start timestamp
     scale:
         const value to fill
-    period:
-        data frequency -- x[i+period] = x[i]
     n_segments:
         number of segments
     freq:
@@ -131,6 +158,11 @@ def generate_const_df(
         scale of added noise
     random_seed:
         random seed
+
+    Raises
+    ------
+    ValueError:
+        Non-integer timestamp parameter is used for integer-indexed timestamp.
     """
     patterns = [[scale] for _ in range(n_segments)]
     df = generate_from_patterns_df(
@@ -147,9 +179,9 @@ def generate_const_df(
 
 def generate_from_patterns_df(
     periods: int,
-    start_time: str,
+    start_time: Optional[Union[str, TimestampType]],
     patterns: List[List[float]],
-    freq: str = "1D",
+    freq: Optional[str] = "D",
     add_noise=False,
     sigma: float = 1,
     random_seed: int = 1,
@@ -173,6 +205,11 @@ def generate_from_patterns_df(
         scale of added noise
     random_seed:
         random seed
+
+    Raises
+    ------
+    ValueError:
+        Non-integer timestamp parameter is used for integer-indexed timestamp.
     """
     n_segments = len(patterns)
     if add_noise:
@@ -183,7 +220,7 @@ def generate_from_patterns_df(
     for idx, pattern in enumerate(patterns):
         samples[idx, :] += np.array(pattern * (periods // len(pattern) + 1))[:periods]
     df = pd.DataFrame(data=samples.T, columns=[f"segment_{i}" for i in range(n_segments)])
-    df["timestamp"] = pd.date_range(start=start_time, freq=freq, periods=periods)
+    df["timestamp"] = _create_timestamp(start_time=start_time, freq=freq, periods=periods)
     df = df.melt(id_vars=["timestamp"], value_name="target", var_name="segment")
     return df
 
@@ -191,8 +228,8 @@ def generate_from_patterns_df(
 def generate_hierarchical_df(
     periods: int,
     n_segments: List[int],
-    freq: str = "D",
-    start_time: str = "2000-01-01",
+    freq: Optional[str] = "D",
+    start_time: Optional[Union[str, TimestampType]] = None,
     ar_coef: Optional[list] = None,
     sigma: float = 1,
     random_seed: int = 1,
@@ -235,6 +272,8 @@ def generate_hierarchical_df(
         ``n_segments`` contains not positive integers
     ValueError:
         ``n_segments`` represents not non-decreasing sequence
+    ValueError:
+        Non-integer timestamp parameter is used for integer-indexed timestamp.
     """
     if len(n_segments) == 0:
         raise ValueError("`n_segments` should contain at least one positive integer!")